Saturday, March 30, 2024 at 10:36 AM West Africa Standard Time

+++ src/fee/transaction-fee/transaction-fee.service.ts
@@ -23,51 +23,42 @@ export class TransactionFeeService {
   }
 
   async createTransactionFee(
-    // userId: ObjectId,
+    userId: ObjectId,
     transactionFeeData: CreateTransactionFeeDTO,
   ): Promise<TransactionFee> {
-    // Check if the user is an admin
-    // await this.checkAdminPermission(userId);
+    await this.checkAdminPermission(userId);
 
-    // Create the transaction fee
     const createdTransactionFee =
       await this.transactionFeeModel.create(transactionFeeData);
 
-    // Return the created transaction fee(s)
     return createdTransactionFee;
   }
 
   async updateTransactionFee(
-    // userId: ObjectId,
+    userId: ObjectId,
     feeId: ObjectId,
     updateData: UpdateTransactionFeeDTO,
   ): Promise<TransactionFee> {
-    // Check if the user is an admin
-    // await this.checkAdminPermission(userId);
+    await this.checkAdminPermission(userId);
 
-    // Find the transaction fee by ID
     const existingFee = await this.transactionFeeModel.findById(feeId).exec();
 
     if (!existingFee) {
       throw new Error('Transaction fee not found');
     }
 
-    // Update the transaction fee with the new data
     Object.assign(existingFee, updateData);
     const updatedFee = await existingFee.save();
 
-    // Return the updated transaction fee
     return updatedFee;
   }
 
   async deleteTransactionFee(
-    // userId: ObjectId,
+    userId: ObjectId,
     feeId: ObjectId,
   ): Promise<boolean> {
-    // Check if the user is an admin
-    // await this.checkAdminPermission(userId);
+    await this.checkAdminPermission(userId);
 
-    // Find the transaction fee by ID and delete it
     const result = await this.transactionFeeModel
       .deleteOne({ _id: feeId })
       .exec();
@@ -80,11 +71,9 @@ export class TransactionFeeService {
   }
 
   private async checkAdminPermission(userId: ObjectId): Promise<void> {
-    // Find the user by ID
     const user = await this.userModel.findById(userId).exec();
 
-    // Check if user exists and is an admin
-    if (!user || user.account_type !== 'admin') {
+    if (!user || user.role !== 'superadmin') {
       throw new Error(
         'User not found or not authorized to manage transaction fees',
       );
